Bugfixes, Makefile, README, demo and test suites

+++ src/prime.h
@@ -1,5 +1,6 @@
 #ifndef PRIME_H
 #define PRIME_H
-int is_prime(const int x);
-int next_prime(int x);
+#include <stddef.h>
+int    is_prime(const int x);
+size_t next_prime(size_t x);
 #endif
